implementation des distances en fonctions des distances

+++ src/model/Titanic.java
@@ -36,7 +36,7 @@ public class Titanic implements IPoint {
 	public int getPclass() {return Pclass;}
 	public String getName() {return name;}
 	public Sexe getSex() {return sex;}
-	public int getAge() {return age;}
+	public double getAge() {return age;}
 	public int getSibSp() {return sibSp;}
 	public int getParch() {return parch;}
 	public String getTicket() {return ticket;}
